requirements.txt
absl-py==1.0.0
attrs==21.2.0
importlib-metadata==4.8.3
iniconfig==1.1.1
numpy==1.19.5
ortools==9.2.9972
packaging==21.3
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
six==1.16.0
toml==0.10.2
typing_extensions==4.0.1
zipp==3.6.0
